Move export assignment methods and removeDefaultExport to ModuledNode.
