+++ src/tests/compiler/file/sourceFileTests.ts
@@ -473,129 +473,6 @@ describe(nameof(SourceFile), () => {
         });
     });
 
-    describe(nameof<SourceFile>(n => n.insertExportAssignments), () => {
-        function doTest(startCode: string, index: number, structures: ExportAssignmentStructure[], expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.insertExportAssignments(index, structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should insert the different kinds of exports", () => {
-            doTest("", 0, [
-                { expression: "5" },
-                { isExportEquals: true, expression: writer => writer.write("6") },
-                { isExportEquals: false, expression: "name" }
-            ], [
-                `export = 5;`,
-                `export = 6;`,
-                `export default name;`
-            ].join("\n") + "\n");
-        });
-
-        it("should insert at the beginning", () => {
-            doTest(`export class Class {}\n`, 0, [{ expression: "5" }], `export = 5;\n\nexport class Class {}\n`);
-        });
-
-        it("should insert in the middle", () => {
-            doTest(`export * from "./file1";\nexport = 6;\n`, 1, [{ expression: "5" }],
-                `export * from "./file1";\n\nexport = 5;\nexport = 6;\n`);
-        });
-
-        it("should insert at the end", () => {
-            doTest(`export class Class {}\n`, 1, [{ expression: "5" }], `export class Class {}\n\nexport = 5;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.insertExportAssignment), () => {
-        function doTest(startCode: string, index: number, structure: ExportAssignmentStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.insertExportAssignment(index, structure);
-            expect(result).to.be.instanceOf(ExportAssignment);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should insert at the specified position", () => {
-            doTest(`export * from "./file1";\nexport = 6;\n`, 1, { expression: "5" },
-                `export * from "./file1";\n\nexport = 5;\nexport = 6;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addExportAssignment), () => {
-        function doTest(startCode: string, structure: ExportAssignmentStructure, expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addExportAssignment(structure);
-            expect(result).to.be.instanceOf(ExportAssignment);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should always add at the end of the file", () => {
-            doTest(`export class MyClass {}\n`, { expression: "5" },
-                `export class MyClass {}\n\nexport = 5;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.addExportAssignments), () => {
-        function doTest(startCode: string, structures: ExportAssignmentStructure[], expectedCode: string) {
-            const { sourceFile } = getInfoFromText(startCode);
-            const result = sourceFile.addExportAssignments(structures);
-            expect(result.length).to.equal(structures.length);
-            expect(sourceFile.getText()).to.equal(expectedCode);
-        }
-
-        it("should add multiple", () => {
-            doTest(`export class MyClass {}\n`, [{ expression: "5" }, { expression: "6" }],
-                `export class MyClass {}\n\nexport = 5;\nexport = 6;\n`);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportAssignments), () => {
-        it("should get the export declarations", () => {
-            const { sourceFile } = getInfoFromText("export = 5; export = 6;");
-            expect(sourceFile.getExportAssignments().length).to.equal(2);
-            expect(sourceFile.getExportAssignments()[0]).to.be.instanceOf(ExportAssignment);
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportAssignment), () => {
-        it("should get the export declaration", () => {
-            const { sourceFile } = getInfoFromText("export = 5; export default 6;");
-            expect(sourceFile.getExportAssignment(e => !e.isExportEquals())!.getText()).to.equal("export default 6;");
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.getExportAssignmentOrThrow), () => {
-        it("should get the export declaration", () => {
-            const { sourceFile } = getInfoFromText("export = 5; export default 6;");
-            expect(sourceFile.getExportAssignmentOrThrow(e => !e.isExportEquals()).getText()).to.equal("export default 6;");
-        });
-
-        it("should throw when not exists", () => {
-            const { sourceFile } = getInfoFromText("");
-            expect(() => sourceFile.getExportAssignmentOrThrow(e => false)).to.throw();
-        });
-    });
-
-    describe(nameof<SourceFile>(n => n.removeDefaultExport), () => {
-        it("should do nothing when there's no default export", () => {
-            const { sourceFile } = getInfoFromText("");
-            sourceFile.removeDefaultExport();
-            expect(sourceFile.getFullText()).to.equal("");
-        });
-
-        it("should remove the default export symbol when one exists", () => {
-            const { sourceFile } = getInfoFromText("export default class Identifier {}");
-            sourceFile.removeDefaultExport();
-            expect(sourceFile.getFullText()).to.equal("class Identifier {}");
-        });
-
-        it("should remove the default export symbol when default exported on a separate statement", () => {
-            const { sourceFile } = getInfoFromText("namespace Identifier {}\nclass Identifier {}\nexport default Identifier;\n");
-            sourceFile.removeDefaultExport();
-            expect(sourceFile.getFullText()).to.equal("namespace Identifier {}\nclass Identifier {}\n");
-        });
-    });
-
     describe(nameof<SourceFile>(n => n.getLanguageVariant), () => {
         it("should return standard when in a ts file", () => {
             const { sourceFile } = getInfoFromText("");
